+++ spec/app/models/pet_owner.rb
@@ -1,6 +1,6 @@
 class PetOwner
   include Mongoid::Document
   field :title
-  embeds_one :pet
+  embeds_one :pet, cascade_callbacks: true
   embeds_one :address, as: :addressable
 end