+++ CHANGELOG.md
@@ -799,6 +799,9 @@ For instructions on upgrading to newer versions, visit
 
 * \#2119 Allow `Criteria#update_all` to accept atomic ops and normal sets.
 
+* \#2117 Ensure embeds one relations have callbacks run when using nested
+  attributes.
+
 * \#2110 `Model#touch` now works properly on embedded documents.
 
 * \#2100 Allow atomic operations to properly execute on paranoid documents